+++ Version_1.1_Final_German/webportal/modules/profile.module.php
@@ -20,19 +20,14 @@
 # notice that explains how it can be used.     #
 ################################################
 
-$std_sid = "show";
-if(isset($_GET['sid']) && $_GET['sid'] != "NULL" && $_GET['sid'] != "" && $_GET['sid'] != "0" && $_GET['sid'] != "false") {
-	$sid = htmlspecialchars($_GET['sid'], ENT_QUOTES);
-} else {
-	$sid = $std_sid;
-}
 class profile {
 	function title() {
 		return "Profil";
 	}
 	function main() {
-		global $id,$sid;
 		$main = new main();
+		$id = $main->id();
+		$sid = $main->sid();
 		$content = "";
 		if(isset($_SESSION['loggedin']) && $_SESSION['loggedin'] == "1") {
 			if($sid == "post") {