+++ Version_1.1_Final_German/webportal/modules/search.module.php
@@ -25,7 +25,6 @@ class search {
 		return "Suchen";
 	}
 	function main() {
-		global $id,$sid;
 		$main = new main();
 		$gbook = "";
 		$blog = "";